+++ CHANGELOG.md
@@ -68,6 +68,7 @@ Developers may need to change their code
 
 ### Minor features
 
+* Added checks of changed handlers or blocked signal after plugin function calls
 * Added set/get pointer API to clixon_data:
    * clicon_ptr_get(), clicon_ptr_set(), 
 * Restconf YANG PATCH according to RFC 8072